requirements.txt
numpy>=1.26
pandas>=2.0
numba>=0.59
websockets>=12.0
pyyaml>=6.0
streamlit>=1.36
tqdm>=4.66
